ci: modifying unit testings for list visualization
