+++ tests/index.test.ts
@@ -1,0 +1,5 @@
+import role from "./Role";
+import statement from "./Statement";
+
+role();
+statement();
